api/requirements.txt
flask
pandas
scikit-learn
scipy
numpy
anthropic
python-dotenv